Fixes the javadoc generation crash.
Was related to inner classes that were crashing an inner
working of sun jdk class.
