+++ pom.xml
@@ -474,7 +474,7 @@
 				<artifactId>maven-javadoc-plugin</artifactId>
 				<version>2.10.3</version>
 				<configuration>
-					<failOnError>true</failOnError>
+					<failOnError>false</failOnError>
 					<sourcepath>target/generated-sources/delombok</sourcepath>
 				</configuration>
 				<reportSets>
