+++ src/main/java/me/moodcat/backend/rooms/package-info.java
@@ -1,0 +1,6 @@
+/**
+ * Contains that the {@link me.moodcat.backend.rooms.RoomBackend}
+ * that uses instances like {@link me.moodcat.backend.rooms.RoomInstance}
+ * and {@link me.moodcat.backend.rooms.SongInstance}.
+ */
+package me.moodcat.backend.rooms;
